+++ release-notes.md
@@ -1,3 +1,15 @@
+# 0.0.3
+
+- [X] Update an existing TOC
+  - [X] Create a section around TOC
+  - [X] Find if a TOC is already present in buffer
+  - [X] Remove old TOC
+  - [X] Create new one
+- [X] Reference markdown-toc version
+- [X] Reference the inspiration from doctoc
+- [X] Upgrade version
+- [X] Release notes
+
 # 0.0.2
 
 - [X] Prepare filesystem to deploy to melpa (avoid having tests files with the file to package)